requirements.txt
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-ckeditor==5.4.0
django-datetimepicker==3.14
django-debug-toolbar==1.9.1
django-js-asset==1.0.0
django-mptt==0.9.0
django-phonenumber-field==2.0.0
django-widget-tweaks==1.4.1
facebookads==2.11.4
idna==2.6
MonthDelta==0.9.1
phonenumberslite==8.9.0
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
urllib3==1.22
